Move custom param converter and related logic from microprofile/metrics to microprofile/tests/tck/tck-metrics so it cannot even accidentally be used by developers
